+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Added
 - LayerTree: added a new parameter at layer `filterRefId` as an index of a configured filter Layer.
+- CompareFeatures: applied the functionality from version 2 to 3 and changed the data display to use shared table (TableComponent).
 
 ### Changed
 
@@ -22,6 +23,7 @@
 - Issue #1314: Added WMS are added in a folder structure in the layerSelection.
 - Issue #1321 Grouped Layer: Child layer adopt the transparency from the config.json.
 - Print-Module: Exchanged deprecated mapfishServiceId for printServiceId and added optional chaining for payload.
+- Routing: Fixed URL of ORS request in TSR
 
 ---